pyproject.toml
[project]
name = "repl-nix-workspace"
version = "1.4.1"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-limiter>=4.0.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "jdatetime>=5.2.0",
    "pillow>=12.0.0",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "urllib3>=2.5.0",
    "werkzeug>=3.1.4",
]

[tool.setuptools]
py-modules = ["app", "main"]

